created Main.java and Train.java
